Add util for building proptypes for withApiFetch data
A lot of the components in this repo retrieve data from HM Stack using
the withApiFetch HOC, which can either return the data or a WP_Error
indicating failure. This adds a helper function in the utils directory
which builds the propTypes for this data field.

Note: This isn't necessariy hugely helpful in itself, but I was looking
forward to some related helper methods related to error handling: for
example, not caching WP_Error responses in localStorage, or displaying a
notice in the component rather than just not displaying anything. Those
kinds of helpers can live in the utils/wp-error module.